WIP: Refactor backend class
Up to now, this was carefully constructed dictionaries being passed
around. Now they are classes and objects, providing a more rigid and
true object-oriented design for current and future backend
implementations, as well as improving documentation and code
readability.

The old concept of passing an "internal ID" was dropped, as the get/set
functions are grouped together with support for individual refreshing.
This improves performance as data can be cached without having to reload
via get_device().

This makes things easier to develop and maintain, especially if there's
new features or changes without bearing the risk of unexpected
breakages.

Due to the structural change, many areas were refactored, including
tests and interfaces.